sma: initial Storage Management Agent structures
Storage Management Agent is a gRPC server that provides an abstraction
layer above the SPDK RPC interface.  The interface aims to provide a set
of methods for managing various protocols (e.g. NVMe, virtio-blk) while
hiding the details of a particular transport.

The external API is defined by `lib/python/spdk/sma/proto/sma.proto`
protobuf file.  It defines the generic gRPC service methods.  The
subsystem-specific parameters are wrapped by a google.protobuf.Any
object and are defined by a subsystem-specific protobuf file.

This patch implements a gRPC service that delegates the work to a specific
subsystem.  A subsystem is a class that implements some of the methods
defined by the service for a given type of devices (e.g. NVMe/TCP,
NVMe/PCIe, virtio-blk, etc.).  For now, only the RPC for creating a
device is implemented, others are added in subsequent patches.

The series implements the generic calls as well as their NVMe/TCP
implementation.  Support for other devce types could be easily added by
creating a new subsystem and defining its protobuf parameter definition.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I17cde3b31d3514878f1027cfcd112b48848f6123