[RFC] sma: nvmf/vfiouser create device implementation
Implementation of VFIOUSER subsystem create device
method with some mandatory sidecar methods. Both socket
full path creation (traddr) and QMP device ID are based
on host_id, PF_ID and VF_ID created in a wat that can
be easly reversed. For QMP protocol methods were added
for probing QMP if device with providen name already
exists or not. It is based on return checking if error
type of device not found occured on QMP.

Signed-off-by: Milosz Linkiewicz <milosz.linkiewicz@intel.com>
Change-Id: I82ed5b573abc595e1c25920a2e04f9a342172e97