[RFC] NVMeOF/VFioUser: Delete device implementation
Implementation of VFIOUSER subsystem delete device method. It uses sleep
to wait for the delete to complete but this is just a walkaround. The QMP
sends device deleted event after completion so ideal implementation
should wait for such to occur before deleting the socket. Method for
socket path clean-up was added.

Signed-off-by: Milosz Linkiewicz <milosz.linkiewicz@intel.com>
Change-Id: Ifb4f6fc9f81618bb10a9d79086b45255fcdf1596