sma: register subsystems based on config file
This allows a user to specify the subsystems to enable.  Those not
specified in the config, won't get initialized and will not service
user's requests.

Additionally, each subsystem can now receive its own configuration in
its init() method.

The subsystem configuration is structured as a list of objects with two
properties: "name" and "params".  The former identifies a subsystem to
enable, while the latter contains a set of options (if any) specific to
that subsystem.  For instance:

```
subsystems:
  - name: 'nvmf-tcp'
    params:
      max_queue_depth: 256
      io_unit_size: 8192
  - name: 'nvmf-vfiouser'
```

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I967016502ad93c243b3a7af58992bde14c44953c