sma: register devices based on config file
This allows a user to specify the types of devices to support.  Those
not specified in the config, won't get initialized and will not service
user's requests.

Additionally, each device manager can now receive its own configuration
in its init() method.

The device configuration is structured as a list of objects with two
properties: "name" and "params".  The former identifies a device to
enable, while the latter contains a set of options (if any) specific to
that device manager.  For instance:

```
devices:
  - name: 'nvmf-tcp'
    params:
      max_queue_depth: 256
      io_unit_size: 8192
  - name: 'nvmf-vfiouser'
```

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I967016502ad93c243b3a7af58992bde14c44953c